--story=874848731 修复删除报错问题

+++ src/backend/maven/biz-maven/src/main/kotlin/com/tencent/bkrepo/maven/service/MavenMetadataService.kt
@@ -9,9 +9,6 @@ import com.tencent.bkrepo.maven.pojo.MavenVersion
 import com.tencent.bkrepo.maven.util.MavenStringUtils.resolverName
 import com.tencent.bkrepo.repository.pojo.node.NodeDetail
 import com.tencent.bkrepo.repository.pojo.node.service.NodeCreateRequest
-import java.time.ZoneId
-import java.time.ZonedDateTime
-import java.time.format.DateTimeFormatter
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.data.mongodb.core.FindAndModifyOptions
@@ -19,6 +16,9 @@ import org.springframework.data.mongodb.core.query.Criteria
 import org.springframework.data.mongodb.core.query.Query
 import org.springframework.data.mongodb.core.query.Update
 import org.springframework.stereotype.Service
+import java.time.ZoneId
+import java.time.ZonedDateTime
+import java.time.format.DateTimeFormatter
 
 @Service
 class MavenMetadataService(
@@ -95,16 +95,20 @@ class MavenMetadataService(
 
     fun delete(mavenArtifactInfo: ArtifactInfo, node: NodeDetail? = null, mavenGavc: MavenGAVC? = null) {
         node?.let {
-            val (criteria, _) = nodeCriteria(
+            val (criteria, mavenVersion) = nodeCriteria(
                 projectId = node.projectId,
                 repoName = node.repoName,
                 metadata = node.metadata,
                 fullPath = node.fullPath
             )
-            criteria?.let {
-                val query = Query(criteria)
-                mavenMetadataDao.remove(query)
+            if (criteria == null) return
+            criteria.apply {
+                mavenVersion?.timestamp?.let {
+                    this.and(TMavenMetadataRecord::timestamp.name).`is`(mavenVersion.timestamp)
+                }
             }
+            val query = Query(criteria)
+            mavenMetadataDao.remove(query)
         }
         mavenGavc?.let {
             val groupId = mavenGavc.groupId
